pyproject.toml
[tool.poetry]
name = "ragpon"
version = "0.1.0"
description = ""
authors = ["tyaso777 <49575436+tyaso777@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
spacy = "^3.7.5"
python-dotenv = "^1.0.1"
sudachipy = "^0.6.8"
peewee = "^3.17.5"
pandas = "^2.2.2"
sudachidict-full = "^20240409"
sudachidict-core = "^20240409"
chromadb = "^0.5.0"
transformers = "^4.41.2"
torch = "^2.3.1"
chardet = "^5.2.0"
pypdf = "^4.2.0"
ja-ginza = "^5.2.0"
pymupdf = "^1.24.5"
truststore = "^0.10.0"
pytest-mock = "^3.14.0"
sentence-transformers = "^3.3.1"
fugashi = "^1.4.0"
unidic-lite = "^1.0.8"
python-docx = "^1.1.2"
pip-audit = "^2.7.3"
pip-licenses = "^5.0.0"
numpy = "^1.24"

[tool.poetry.group.dev.dependencies]
notebook = "^7.2.1"
ipykernel = "^6.29.4"
black = "^24.4.2"
flake8 = "^7.0.0"
mypy = "^1.10.0"
isort = "^5.13.2"
pytest = "^8.2.2"
sphinx = "^8.1.3"
pylint = "^3.3.3"
sphinx-autobuild = "^2024.10.3"
sphinx-rtd-theme = "^3.0.2"
graphviz = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
